% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_img1}
\alias{demo_img1}
\title{demo_img1}
\format{A raw vector of length 1114967.}
\source{
\href{https://unsplash.com/photos/yQILyG_fGuE}{Unsplash user Szabo Viktor}.
}
\usage{
demo_img1
}
\description{
Demo image 1.
}
\keyword{datasets}
