% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_img2}
\alias{demo_img2}
\title{demo_img2}
\format{A raw vector of length 385460.}
\source{
\href{https://unsplash.com/photos/LWuHsMPzoss}{Unsplash user Ilona Panych}.
}
\usage{
demo_img2
}
\description{
Demo image 2.
}
\keyword{datasets}
