% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_img3}
\alias{demo_img3}
\title{demo_img3}
\format{A raw vector of length 233103.}
\source{
\href{https://unsplash.com/photos/i0yjni9irCA}{Unsplash user Mike Von}.
}
\usage{
demo_img3
}
\description{
Demo image 3.
}
\keyword{datasets}
