% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glitch_functions.R
\name{glitch_far}
\alias{glitch_far}
\title{Glitch using find and replace.}
\usage{
glitch_far(input_data, n_changes = 5, tune = 100)
}
\arguments{
\item{input_data}{A raw vector representing the file to databend.}

\item{n_changes}{A numeric value. How many unique elements \emph{F} should be
selected for replacement?}

\item{tune}{A numeric value between 1 and 100. What percentage of the
occurences of \eqn{F} should be replaced?}
}
\value{
A raw vector representing a modified version of the input vector.
}
\description{
A function for databending a raw vector by randomly sampling unique observed
values \eqn{F} and replacing them with another randomly sampled unique value
\eqn{R}.
}
\examples{
data(demo_img1)
my_glitch <- glitch_far(demo_img1, n_changes = 3, tune = 5)
\dontrun{
my_glitch \%>\%
  magick::image_read() \%>\%
  magick::image_scale(geometry_size_pixels(width = 550)) \%>\%
  print()
}

}
\concept{glitch functions}
