% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glitchr.R
\docType{package}
\name{glitchr}
\alias{glitchr}
\title{glitchr: A package for databending raw vectors in R.}
\description{
The glitchr package provides some functions for databending in R. The
functions take raw vectors and manipulate them in various ways to produce
glitched versions of the original files.
}
\section{Glitch functions}{

\itemize{
  \item glitch_far()
  \item glitch_co()
}
}

