% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2jpeg.R
\name{jpeg2jpeg}
\alias{jpeg2jpeg}
\title{Convert JPEG to 6 MP JPEG}
\usage{
jpeg2jpeg(dir = getwd())
}
\arguments{
\item{dir}{A directory of *[jJ][pP][eE]?[gG] files.}
}
\description{
If an image is a JPEG, make it a reasonably sized JPEG.
}
