% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_vector}
\alias{plot_vector}
\title{Plot the raw vector as an image in the RStudio Viewer.}
\usage{
plot_vector(raw_vector, height_px = NULL, width_px = 550)
}
\arguments{
\item{raw_vector}{The raw vector object in R which you wish to plot.}

\item{height_px}{The width in pixels to scale the image before viewing. To
ensure the correct aspect ratio is maintained when specifying height, set
\code{width_px = NULL}.}

\item{width_px}{The width in pixels to scale the image before viewing.}
}
\value{
A printable magick tibble.
}
\description{
Helper function to plot the raw vector as an image in the RStudio Viewer pane
(or another configured browser) using the magick package.
}
\examples{
\dontrun{
  glitch_it(demo_img1, method = "far",
            n_changes = 3, tune = 5) \%>\%
  plot_vector(width_px = 300)
}
}
