% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square Image}
\usage{
square(img)
}
\arguments{
\item{img}{An image with any number of color channels.}
}
\description{
If an image is not square, add white space to make it square.
}
