% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super8.R
\name{super8_18fps}
\alias{super8_18fps}
\title{Adjust Frame Rate of Super8 Scans}
\usage{
super8_18fps(dir = getwd(), negate = TRUE)
}
\arguments{
\item{dir}{Directory}

\item{negate}{A logical. Toggles whether to negate images.}
}
\description{
Adjust Frame Rate of Super8 Scans
}
