% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2jpeg.R
\name{tiff2jpeg}
\alias{tiff2jpeg}
\title{Convert TIFF to 6 MP JPEG}
\usage{
tiff2jpeg(dir = getwd())
}
\arguments{
\item{dir}{A directory of *[tT][iI][fF][fF]? files.}
}
\description{
If an image is a TIFF, make it a reasonably sized JPEG.
}
