% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v600.R
\name{v600_120}
\alias{v600_120}
\title{Crop Scans from V600}
\usage{
v600_120(dir = getwd(), negate = TRUE, split = 1)
}
\arguments{
\item{dir}{Directory}

\item{negate}{A logical. Toggles whether to negate images.}

\item{split}{An integer. Number of frames per scan.}
}
\description{
Crop Scans from V600
}
