% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v600.R
\name{v600_4x5}
\alias{v600_4x5}
\title{Crop Scans from V600}
\usage{
v600_4x5(dir = getwd(), negate = TRUE)
}
\arguments{
\item{dir}{Directory}

\item{negate}{A logical. Toggles whether to negate images.}
}
\description{
Crop Scans from V600
}
