% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{squaregram}
\alias{squaregram}
\title{Square Image}
\usage{
squaregram(dir = getwd())
}
\arguments{
\item{img}{A directory of .jpe?g files.}
}
\description{
If an image is not square, add white space to make it square.
}
